Refactor: move RandString functions to rand.go
