+++ rand.go
@@ -2,6 +2,13 @@ package gofn
 
 import "math/rand"
 
+var (
+	StrLowerAlpha   = []rune("abcdefghijklmnopqrstuvwxyz")
+	StrUpperAlpha   = []rune("ABCDEFGHIJKLMNOPQRSTUVWXYZ")
+	StrDigits       = []rune("0123456789")
+	StrDefaultChars = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789")
+)
+
 // RandChoiceMaker a struct for picking up items randomly from a list of items
 type RandChoiceMaker[T any] struct {
 	source   []*T // Use pointers to slice items to gain more performance when item type is struct
@@ -72,4 +79,19 @@ func Shuffle[T any](s []T, randFuncs ...func(n int) int) []T {
 		result = append(result, item)
 	}
 	return result
+}
+
+// RandString generates a random string
+func RandString(n int) string {
+	return RandStringEx(n, StrDefaultChars)
+}
+
+// RandStringEx generates a random string
+func RandStringEx(n int, allowedChars []rune) string {
+	b := make([]rune, n)
+	numChars := len(allowedChars)
+	for i := range b {
+		b[i] = allowedChars[rand.Intn(numChars)] // nolint: gosec
+	}
+	return string(b)
 }
